requirements.txt
django
gunicorn
django-heroku
mysql-connector
dialogflow
requests